231009
dd

+++ Assets/Scripts/Status.cs
@@ -6,6 +6,6 @@ public class Status : MonoBehaviour
 {
     public int Atk = 0;
     public float Spd = 0;
-    public int CurHp = 0;
-    public int MaxHp = 0;
+    public float CurHp = 0;
+    public float MaxHp = 0;
 }
